refactor: create a lib folder, move Sanity client into it.
